refactor: state내 함수들 분리
