+++ src/functions/add-submit-event.js
@@ -1,12 +1,13 @@
 import { state } from '../state/index.js';
 import { USER_INPUT, RESULT_DIV } from '../constants/html-doms.js';
-import { isUserInputValid } from '../functions/user-input.js';
+import { isUserInputValid, setUserInput } from '../functions/user-input.js';
+import { initBallStrike } from './util.js';
 import { makeResultDom } from '../functions/game-result.js';
 
 export function onSubmitClick(event, play) {
   event.preventDefault();
-  state.setUserInput(parseInt(USER_INPUT.value));
-  state.initBallStrike();
+  setUserInput(parseInt(USER_INPUT.value));
+  initBallStrike();
   if (isUserInputValid(USER_INPUT.value)) {
     const resultMessage = play(state.computerInput, state.userInput);
     makeResultDom(resultMessage);
