+++ src/index.js
@@ -1,13 +1,12 @@
-import { state } from './state/index.js';
+import { setComputerInput, setBallStrike } from './functions/util.js';
 import { SUBMIT_BUTTON } from './constants/html-doms.js';
 import { makeResultMessage } from './functions/game-result.js';
-
 import { onSubmitClick } from './functions/add-submit-event.js';
 
 export default function BaseballGame() {
-  state.setComputerInput();
+  setComputerInput();
   this.play = function (computerInput, userInput) {
-    state.setBallStrike(computerInput, userInput);
+    setBallStrike(computerInput, userInput);
 
     return makeResultMessage();
   };
